Candidate filter policy for Tor Browser in Oz-Daemon
This is a filter policy for oz-daemon, though it is meant to for Tor
Browser. The assumption is that the only application in Oz using the Tor
Control service is TB for now. This is going to change, and we wil need
to support better application multiplexing in Roflcptor before we can
handle more than one Oz application with unique policies. Please test
before including it. Special attention was made to policy creation to
support features such as the circuit geo-ip dialog.